Basic cutscene camera movement
Updated the CameraTracker script to LERP the camera to a target location during a cutscene.

For each point the camera moves to, a "CamCutscnMov" needs to be added to the scene, then it needs to be added to the timeline with the relevant signals added.

The CamCutscnMov has a script on it which automatically grabs the first camera in the scene and makes a reference to it to control it.